renaming flutter project and updating flutter packages
