+++ .metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: "abb292a07e20d696c4568099f918f6c5f330e6b0"
+  revision: "ba393198430278b6595976de84fe170f553cc728"
   channel: "stable"
 
 project_type: app
@@ -13,26 +13,26 @@ project_type: app
 migration:
   platforms:
     - platform: root
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: android
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: ios
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: linux
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: macos
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: web
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
     - platform: windows
-      create_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
-      base_revision: abb292a07e20d696c4568099f918f6c5f330e6b0
+      create_revision: ba393198430278b6595976de84fe170f553cc728
+      base_revision: ba393198430278b6595976de84fe170f553cc728
 
   # User provided section
