+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application
-        android:label="gui_flutter"
+        android:label="medipal"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
