+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: gui_flutter
+name: medipal
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: '>=3.3.1 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,10 +35,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  firebase_core: ^2.27.0
-  firebase_analytics: ^10.8.9
-  cloud_firestore: ^4.15.8
-  firebase_database: ^10.4.9
+  firebase_core: ^2.27.1
+  firebase_analytics: ^10.8.10
+  firebase_auth: ^4.17.9
+  cloud_firestore: ^4.15.9
+  firebase_database: ^10.4.10
 
 dev_dependencies:
   flutter_test:
